requirements.txt
opencv-python==4.8.1.78
numpy==1.24.3
smbus2==0.4.3
filterpy==1.4.2
RPi.GPIO==0.7.0
